Check win and exit TicTacToe
Game field displayed in a separate function

Func game_core shows the game field two times (at the start
and in a while cycle). I decided it is more correct to follow DRY.

Determine the winner and finish game

Winner is checked in such way so that other features of the game
can be added in the future. Such as game field size and shape,
required number of characters in a row to win and enabling/disabling
the boundaries of game field; Also, the game ends if there is a winner
or if enter the "Выход" command;

Cleanup console

It works only in the terminal if launching ".py " file.